Created a new 'lib' directory and moved util to it.
