+++ lib/util.c
